# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wallet",
        "contact": {
            "email": "zoheirmaz.zm@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/credit/charge": {
            "post": {
                "tags": [
                    "Credit"
                ],
                "summary": "Charge user credit",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "integer"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "register sucessful"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            }
        },
        "/credit/amount": {
            "get": {
                "tags": [
                    "Credit"
                ],
                "summary": "Get user credits amount",
                "parameters": [
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "User mobile",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "requst is sucessful"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            }
        },
        "/coupon/apply": {
            "post": {
                "tags": [
                    "Coupon"
                ],
                "summary": "apply a coupon and get credit",
                "parameters": [
                    {
                        "name": "coupon_id",
                        "in": "query",
                        "description": "Coupon id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "User mobile",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "requst is sucessful"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}